backend/requirements.txt
fastapi
uvicorn
pillow
fonttools
python-multipart
fonttools
